refactor: Seperate Input Components and Style from Calendar.tsx
